Add files via upload
Introduction/Business Problem:

How to choose the Best Location to start an Asian Restaurant in London ?

Opening a restaurant is all about location, location, location. However, not every restaurant is suitable for every location, and vice versa. It comes down to a combination of restaurant style, target audience, your competitors. If you can define your restaurant type and identify your target demographic and its most populated areas, you’ll be well on your way to choosing a restaurant location that sets your business up for success.There’s a lot of work, planning and preparation that goes into opening a restaurant. Here are a few tips on how to open a restaurant in London and some steps you’ll need to take in order to achieve your goal.

1.	Restuarant style: The first thing you need to decide is what type of restaurant you want to open and the style it will hold. Having this set in stone will make the other steps a lot easier. To make things a little easier, here are three categories you need to consider:

•	Food: What type of food will you be serving? Will you specialise in a certain cuisine?

•	Service: What type of service will you be providing? Waiter service? Self-service?

•	Atmosphere: What type of atmosphere do you want your establishment to hold? Do you want to open a casual, self-service cafe? Or a fine-dining, high-end restaurant?

2.	Your Target Market:

In order to open an asian restaurant in London, you need to be familiar with exactly who you are aiming to bring into your venue. Are you targeting audience from Indian, Paskistani, Bangladeshhi and Chinese origin ? You need to also be aware of the age group, the location, the amount of money they are willing to spend.

3.	Who are your competitors?

As well as familiarising yourself with your audience, you must do so with your competitors. Look at similar businesses in your area and well-renowned restaurants of the same category. Consider the ways in which they market themselves and the service they provide. What works and what doesn’t? You’ll need to be able to compete with these establishments, so take on board what they’re already doing and better it.

Be aware of the type of your establishment and the number of similar restaurants in the area you are looking to open it. If there is already a lot of them, chances are you’re entering business in an over-saturated area and the potential for you to succeed quickly may be diminished. Consider opening in an area where there is a gap in the
 
market and demand for your service. However, you must also consider the opposing. If there are no restaurants in the area similar to your concept, why? It could simply be that nobody has followed that path yet, or because it will not work. Do your research to ensure you have the right location to suit your restaurant.

4.	Create your menu:

The menu is the core of any restaurant. It’s important you get it right, as it’s the deciding factor for customers deliberating a visit. Your service could be impeccable, decor and atmosphere exquisite, but without a menu to match, you shall have no such luck in success. So when you are targeting the asian ethnicity audience, your menu should include cuisines from India, Pakistan, Bangladesh and China.

5.	Location and Premises:

There are many factors that will contribute to where you decide your premises to be. It can be challenging to find a venue that will factor in all of your conditions, so it is likely you will have to compromise on a few things. However, here are the main factors you should consider when finding premises for your new restaurant:

•	Location: How accessible is the location? Are there many competitors nearby?

•	Building type: What type of building would you like your restaurant to be in?

•	Target market: Make sure it is in an area where your target market is also.

•	Cost: Your ideal location may prove to be a little costly. Make sure that there is a realistic price point.